Make module updates asynchronous
This transforms the blocking module updates into asynchronous workloads
by performing their updates with calloop's timers, together with a
process watchdog.

Calloop's timers are responsible for kicking off the subprocesses
whenever the individual modules should be updated, which are then handed
off to the reaper.

The reaper integrates with calloop through a SIGCHLD listener, updating
its internal list of processes whenever a SIGCHLD signal is received.
The child's output is then awaited and passed to the callback that was
registered with the reaper.

Closes #7.