Add brightness module to drawer
This adds a new brightness slider to the drawer, allowing the readout
and adjustment of screen brightness with udev.

To allow for this extra flexibility with different drawer module types,
the module system has been reworked a bit to more heavily rely on
traits. This simplifies the check for which modules go into which window
by building the API around the `panel_module` and `drawer_module`
functions on the `Module` trait.

Closes #8.