Fix process reaping
Previously subprocesses would only get reaped when a `SIGCHLD` signal
with their PID attached was sent. However it is possible for multiple
`SIGCHLD`s to get bundled at which point the exact PIDs are lost.

To make sure all dead children are reaped whenever a `SIGCHLD` comes in,
this patch uses `try_wait` on every known child and reaps it if it is
dead.

Closes #21.