Fix drawer jumping on first open
This fixes an issue where the drawer would immediately jump to maximized
size the first time it is opened. This happened since the logic added in
74dafa1 detects the panel as already opened on the initial resize from
1x1 to the target size.

To ensure this logic doesn't get in the way while opening/closing the
drawer, the logic is now skipped unless there is no active touch input
for changing the drawer's size.