Add fractional scaling support
This patch adds support for the fractional scaling protocol, fixing an
issue where the drawer would not have the proper scale when opened after
the output scale was changed.

This also fixes an issue with the drawer's EGL surface sizes latching
since they were always created at a fixed size of 1x1 in the `show`
function. Instead the surfaces are now created dynamically with the
correct size in the `resize` function.