Fix modem change tracking
This fixes an issue where modems being added/removed wasn't handled
correctly, since the message in their DBus stream contains the path to
ModemManager instead of the modem being added/removed.

To ensure modems are tracked properly, the full list of modem objects is
now retrieved whenever any modem is added or removed.

This also fixes an issue where the modem state/signal quality streams
were not updated to use the new modem after switching to it.

Additionally it fixes an issue where the removal of every modem would
not reset the modem connectivity object back to the default.