Fix broken drawer textures after scale change
This fixes an issue with the arrow at the bottom of the drawer, where it
would turn into a solid black rectangle after the scale is changed.

This is due to the renderer being resized, invalidating the subtexture
and thus requiring a redraw. To fix this we now redraw the arrow
whenever the scale or size of the window has changed.

Closes #30.