Add volume level as panel background
This adds a new type of UI element for values which are generally in the
range of 0-100% by drawing a bar over the background color relative to
the current value of the module.

A new volume module is introduced which uses this method to indicate the
current volume level when it is actively being changed.

Closes #34.