Improve WiFi strength representation
This changes the WiFi strength SVG conversion to round to the nearest
SVG, rather than ceiling its value. This should make the distinction
between a good and a great connection a little clearer.

Additionally a new SVG for really poor connections has been introduced,
which is used for a connection quality that rounds to 0.