Fix invalid drawer size on startup
This fixes an issue where the drawer size would be invalid when opening
it with a tap (instead of drag) for the first time after startup, since
the drawer was running into the Mesa resize latching bug.

To fix this, resize events are now cached and only applied to the
renderer right before rendering.